Added _add_wave_by_call() code, including function pointer in WaveData struct and (almost) all the necessary handling in the various functions that would care (_add_wave_backend, _draw, etc).
Due to a profound lack of sleep, the code is being checked in in an
uncompilable state; line 500 of gtkwavedraw.c (calling the cb function)
is wrong, and neither my brain nor K&R want to tell me how to fix it.  If
anyone else feels like correcting my moronicism, please do.